Split config_upgrade.py into infrastructure and hooks
Move all upgrade_to_v* functions and their constants to a new
config_upgrade_hooks.py module. config_upgrade.py keeps the
infrastructure: autodetect_upgrade_hooks(), upgrade_config(),
temp_option(), rename_option(), and UpgradeHooksAutodetectError.

The autodetect machinery now scans config_upgrade_hooks by default
instead of its own module.

Defer the UiTheme import in upgrade_to_v3_0_0dev7 to avoid loading
UI modules in headless/CLI contexts.

Split tests to mirror the source layout:
- test_config_upgrade.py: autodetect infrastructure tests
- test_config_upgrade_hooks.py: individual hook tests using
  'import picard.config_upgrade_hooks as hooks'

Adding a new upgrade hook now only requires editing
config_upgrade_hooks.py and test_config_upgrade_hooks.py.
